+++ forui/CHANGELOG.md
@@ -55,6 +55,11 @@
 
 ## 0.21.0
 
+### `FAutocomplete`
+* Add support for generic autocomplete options via `FAutocomplete<T>`.
+* Add `FAutocomplete.displayStringForOption` and `FAutocomplete.onSelect`.
+
+
 ### `FTypography`
 * Add `FTypography.defaultFontFamily` static constant.